+++ metadata.rb
@@ -1,6 +1,6 @@
-name             'dnsupdate-nativex'
-maintainer       'NativeX'
-maintainer_email 'adrian.herrera@nativex.com'
+name             'dnsupdate-simplyadrian'
+maintainer       'simplyadrian'
+maintainer_email 'simplyadrian@gmail.com'
 license          'All rights reserved'
 description      'Update Route53 Zone'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))